reduce line code
