+++ src/components/Pelaporan/FormHealthWorker.vue
@@ -79,7 +79,7 @@
   </div>
 </template>
 <script>
-import { protectionToolOptions } from '@/utils/constantVariable'
+import { protectionToolOptions } from '@/utils/constantOption'
 import { ValidationProvider } from 'vee-validate'
 
 export default {
