+++ src/components/Pelaporan/FormSupportingInvestigation.vue
@@ -415,7 +415,7 @@
   </div>
 </template>
 <script>
-import { protectionToolOptions } from '@/utils/constantVariable'
+import { protectionToolOptions } from '@/utils/constantOption'
 import { ValidationProvider } from 'vee-validate'
 
 export default {
