+++ src/components/Pelaporan/FormTravelHistory.vue
@@ -331,7 +331,7 @@
   </div>
 </template>
 <script>
-import { transportOptions } from '@/utils/constantVariable'
+import { transportOptions } from '@/utils/constantOption'
 import { ValidationProvider } from 'vee-validate'
 import { mapGetters } from 'vuex'
