+++ src/components/Pelaporan/dialogUpdateHistoryCase.vue
@@ -67,13 +67,13 @@
 </template>
 <script>
 import { ValidationObserver } from 'vee-validate'
+import { ResponseRequest } from '@/utils/constantVariable'
 import {
-  ResponseRequest,
   symptomOptions,
-  additionalConditionOptions,
   answerList,
-  yesOrNoAnswer
-} from '@/utils/constantVariable'
+  yesOrNoAnswer,
+  additionalConditionOptions
+} from '@/utils/constantOption'
 import { validateScrollUp } from '@/utils/utilsFunction'
 import { mapGetters } from 'vuex'
 import EventBus from '@/utils/eventBus'
