+++ src/store/modules/hospitalReferral/index.js
@@ -1,0 +1,9 @@
+import actions from './actions'
+
+const state = {}
+
+export default {
+  namespaced: true,
+  state,
+  actions
+}
