+++ src/store/modules/reports/actions.js
@@ -1,5 +1,4 @@
 import requestServer from '@/api'
-import request from '@/utils/request'
 import {
   getRequestDetailHomeAdress
 } from '@/utils/utilsFunction'
@@ -137,9 +136,7 @@ export default {
       return error.response
     }
   },
-  async updateHistoryCase({
-    commit
-  }, body) {
+  async updateHistoryCase({ commit }, body) {
     const { idHistory, data } = body
     try {
       const response = await requestServer(`/api/history_cases/${idHistory}`, 'PUT', data)
@@ -148,64 +145,9 @@ export default {
       return error.response
     }
   },
-  async deleteHistoryCase({
-    commit
-  }, id) {
+  async deleteHistoryCase({ commit }, id) {
     try {
       const response = await requestServer(`/api/history_cases/${id}`, 'DELETE')
-      return response
-    } catch (error) {
-      return error.response
-    }
-  },
-  async exportExcelCase({ commit }, params) {
-    try {
-      const response = await request({
-        url: `/api/cases-export`,
-        method: 'GET',
-        params: params,
-        responseType: 'blob'
-      })
-      return response
-    } catch (error) {
-      return error.response
-    }
-  },
-  async exportExcelHistory({ commit }, params) {
-    try {
-      const response = await request({
-        url: `/api/history-export`,
-        method: 'GET',
-        params: params,
-        responseType: 'blob'
-      })
-      return response
-    } catch (error) {
-      return error.response
-    }
-  },
-  async exportExcelDailyReport({
-    commit
-  }, params) {
-    try {
-      const response = await request({
-        url: `/api/reports/daily-report-xls`,
-        method: 'GET',
-        params: params,
-        responseType: 'blob'
-      })
-      return response
-    } catch (error) {
-      return error.response
-    }
-  },
-  async importExcel({ commit }, data) {
-    const {
-      formData,
-      onProgress
-    } = data
-    try {
-      const response = await requestServer('/api/v2/cases-import', 'UPLOAD', formData, onProgress)
       return response
     } catch (error) {
       return error.response
@@ -227,17 +169,8 @@ export default {
       return error.response
     }
   },
-  async listMedicalFacility({ commit }, params) {
-    try {
-      const response = await requestServer('/api/users-listid', 'GET', params)
-      return response
-    } catch (error) {
-      return error.response
-    }
-  },
   async verifyCase({ commit }, params) {
-    const id = params.id
-    const data = params.data
+    const { id, data } = params
     try {
       const response = await requestServer(`/api/cases/${id}/verifications`, 'POST', data)
       return response
@@ -252,94 +185,6 @@ export default {
       return response
     } catch (e) {
       return e
-    }
-  },
-  async printPEForm({ commit }, id) {
-    try {
-      const response = await request({
-        url: `/api/v2/cases/${id}/export-to-pe-form`,
-        method: 'GET',
-        responseType: 'blob'
-      })
-      return response
-    } catch (error) {
-      return error.response
-    }
-  },
-  async hospitalRefferalNewCase({ commit }, data) {
-    try {
-      const response = await requestServer(`/api/cases-transfer`, 'POST', data)
-      return response
-    } catch (error) {
-      return error.response
-    }
-  },
-  async caseHospitalRefferal({ commit }, params) {
-    const id = params.id
-    const data = params.data
-    try {
-      const response = await requestServer(`/api/cases/${id}/transfers`, 'POST', data)
-      return response
-    } catch (error) {
-      return error.response
-    }
-  },
-  async caseHospitalRefferalRevise({ commit }, params) {
-    const {
-      idCase,
-      idTransfer,
-      data
-    } = params
-    try {
-      const response = await requestServer(`api/cases/${idCase}/transfers/${idTransfer}/revise`, 'POST', data)
-      return response
-    } catch (error) {
-      return error.response
-    }
-  },
-  async caseHospitalReferralInOut({ commit }, data) {
-    const {
-      type,
-      params
-    } = data
-    try {
-      const response = await requestServer(`/api/cases-transfer/${type}`, 'GET', params)
-      return response
-    } catch (error) {
-      return error.response
-    }
-  },
-  async caseHospitalReferralSummary({ commit }, data) {
-    const {
-      type
-    } = data
-    try {
-      const response = await requestServer(`/api/cases-transfer-summary/${type}`, 'GET')
-      return response
-    } catch (error) {
-      return error.response
-    }
-  },
-  async actionHospitalReferral({ commit }, params) {
-    const {
-      idCase,
-      idTransfer,
-      actions,
-      body
-    } = params
-    try {
-      const response = await requestServer(`api/cases/${idCase}/transfers/${idTransfer}/${actions}`, 'POST', body)
-      return response
-    } catch (error) {
-      return error.response
-    }
-  },
-  async caseHospitalReferralHistory({ commit }, id) {
-    try {
-      const response = await requestServer(`/api/cases/${id}/transfers`, 'GET')
-      return response
-    } catch (error) {
-      return error.response
     }
   },
   async revampGetNik({ commit }, params) {
@@ -361,26 +206,6 @@ export default {
   async createRevampReportCase({ commit }, data) {
     try {
       const response = await requestServer('/api/cases-revamp', 'POST', data)
-      return response
-    } catch (error) {
-      return error.response
-    }
-  },
-  async updateCloseContact({ commit }, data) {
-    const id_close_contact = await data.id
-    await delete data['id']
-    try {
-      const response = await requestServer(`/api/cases-revamp/${id_close_contact}/contact`, 'PUT', data.data)
-      return response
-    } catch (error) {
-      return error.response
-    }
-  },
-  async addCloseContact({ commit }, data) {
-    const id_case = await data.id
-    await delete data['id']
-    try {
-      const response = await requestServer(`/api/cases/${id_case}/closecontact`, 'POST', data.data)
       return response
     } catch (error) {
       return error.response
